GEODE-6096: Move VersionManager to org.apache.geode.test.version
