+++ geode-assembly/src/test/java/org/apache/geode/test/junit/rules/GfshRuleTest.java
@@ -23,9 +23,9 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
 
-import org.apache.geode.test.dunit.standalone.VersionManager;
 import org.apache.geode.test.junit.categories.GfshTest;
 import org.apache.geode.test.junit.rules.gfsh.GfshRule;
+import org.apache.geode.test.version.VersionManager;
 
 @Category({GfshTest.class})
 public class GfshRuleTest {
