+++ geode-core/src/distributedTest/java/org/apache/geode/security/ClientAuthenticationPart2DUnitTest.java
@@ -31,9 +31,9 @@ import org.junit.runners.Parameterized;
 import org.apache.geode.internal.cache.tier.MessageType;
 import org.apache.geode.internal.cache.tier.sockets.Message;
 import org.apache.geode.internal.cache.tier.sockets.ServerConnection;
-import org.apache.geode.test.dunit.standalone.VersionManager;
 import org.apache.geode.test.junit.categories.SecurityTest;
 import org.apache.geode.test.junit.runners.CategoryWithParameterizedRunnerFactory;
+import org.apache.geode.test.version.VersionManager;
 
 /**
  * this class contains test methods that used to be in its superclass but that test started taking
