+++ geode-dunit/src/main/java/org/apache/geode/test/dunit/Host.java
@@ -19,7 +19,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 import org.apache.geode.test.dunit.standalone.RemoteDUnitVMIF;
-import org.apache.geode.test.dunit.standalone.VersionManager;
+import org.apache.geode.test.version.VersionManager;
 
 /**
  * This class represents a host on which a remote method may be invoked. It provides access to the
