+++ geode-junit/src/main/java/org/apache/geode/test/version/VersionManager.java
@@ -12,7 +12,7 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package org.apache.geode.test.dunit.standalone;
+package org.apache.geode.test.version;
 
 import java.io.IOException;
 import java.io.InputStream;
