+++ geode-junit/src/test/java/org/apache/geode/test/version/VersionManagerJUnitTest.java
@@ -12,7 +12,7 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package org.apache.geode.test.dunit.standalone;
+package org.apache.geode.test.version;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
