+++ geode-wan/src/upgradeTest/java/org/apache/geode/cache/wan/WANRollingUpgradeCreateGatewaySenderMixedSiteOneCurrentSiteTwo.java
@@ -26,8 +26,8 @@ import org.apache.geode.test.dunit.DistributedTestUtils;
 import org.apache.geode.test.dunit.Host;
 import org.apache.geode.test.dunit.NetworkUtils;
 import org.apache.geode.test.dunit.VM;
-import org.apache.geode.test.dunit.standalone.VersionManager;
 import org.apache.geode.test.junit.rules.GfshCommandRule;
+import org.apache.geode.test.version.VersionManager;
 
 public class WANRollingUpgradeCreateGatewaySenderMixedSiteOneCurrentSiteTwo
     extends WANRollingUpgradeDUnitTest {
